Commiting the modifications that had been made to the trunk version and that were not kept when merging with the api2 fork.

+++ src/fr/ign/cogit/geoxygene/util/algo/MathUtil.java
@@ -1,29 +1,22 @@
 /*
- * This file is part of the GeOxygene project source files.
- * 
- * GeOxygene aims at providing an open framework which implements OGC/ISO
- * specifications for the development and deployment of geographic (GIS)
- * applications. It is a open source contribution of the COGIT laboratory at the
- * Institut Géographique National (the French National Mapping Agency).
- * 
- * See: http://oxygene-project.sourceforge.net
- * 
- * Copyright (C) 2005 Institut Géographique National
- * 
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or any later version.
- * 
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- * 
- * You should have received a copy of the GNU Lesser General Public License
- * along with this library (see file LICENSE if present); if not, write to the
- * Free Software Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA
- * 02111-1307 USA
+ * This file is part of the GeOxygene project source files. GeOxygene aims at
+ * providing an open framework which implements OGC/ISO specifications for the
+ * development and deployment of geographic (GIS) applications. It is a open
+ * source contribution of the COGIT laboratory at the Institut Géographique
+ * National (the French National Mapping Agency). See:
+ * http://oxygene-project.sourceforge.net Copyright (C) 2005 Institut
+ * Géographique National This library is free software; you can redistribute it
+ * and/or modify it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation; either version 2.1 of the License,
+ * or any later version. This library is distributed in the hope that it will be
+ * useful, but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser
+ * General Public License for more details. You should have received a copy of
+ * the GNU Lesser General Public License along with this library (see file
+ * LICENSE if present); if not, write to the Free Software Foundation, Inc., 59
+ * Temple Place, Suite 330, Boston, MA 02111-1307 USA
  */
+
 package fr.ign.cogit.geoxygene.util.algo;
 
 import java.util.List;
@@ -46,7 +39,7 @@ public class MathUtil {
     }
     double somme = 0.0;
     for (Double val : liste) {
-      somme += val;
+      somme += val.doubleValue();
     }
     return somme / liste.size();
   }
@@ -62,7 +55,7 @@ public class MathUtil {
     }
     double somme = 0.0;
     for (Double val : liste) {
-      somme += val * val;
+      somme += val.doubleValue() * val.doubleValue();
     }
     return somme / liste.size();
   }
@@ -114,7 +107,7 @@ public class MathUtil {
     }
     double min = Double.MAX_VALUE;
     for (Double val : liste) {
-      min = Math.min(min, val);
+      min = Math.min(min, val.doubleValue());
     }
     return min;
   }
@@ -130,7 +123,7 @@ public class MathUtil {
     }
     double max = -Double.MAX_VALUE;
     for (Double val : liste) {
-      max = Math.max(max, val);
+      max = Math.max(max, val.doubleValue());
     }
     return max;
   }
@@ -147,7 +140,7 @@ public class MathUtil {
     }
     Double[] listeTriee = liste.toArray(new Double[0]);
     java.util.Arrays.sort(listeTriee);
-    return listeTriee[liste.size() / 2];
+    return listeTriee[liste.size() / 2].doubleValue();
   }
 
   /**
